Moved clients to src/test, refactor of clients.
